requirements.txt
flask
pillow
numpy
scikit-learn
tensorflow
xgboost
imbalanced-learn
opencv-python
